pyproject.toml
[project]
name = "jogy-backend"
version = "0.1.0"
description = "Jogy App Backend - A geo-social application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "geoalchemy2>=0.14.3",
    "alembic>=1.13.1",
    
    # Redis
    "redis>=5.0.1",
    
    # Security
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    
    # Validation & Settings
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    
    # Task Queue (Optional)
    "celery>=5.3.6",
    
    # Utilities
    "httpx>=0.26.0",
    "shapely>=2.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
